+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- [ISSUE-139](https://github.com/dailymotion/tartiflette/issues/139): Handle multiple exceptions on argument coercing.
+
 ## [Released]
 
 ### [0.6.6] - 2019-03-13